pyproject.toml
[tool.poetry]
name = "lambda-ocr-project"
version = "0.1.0"
description = "OCR Lambda function for AWS ECR deployment"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "lambdas"}]

[tool.poetry.dependencies]
python = "~3.11"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^5.0.0"
black = "^24.8.0"
flake8 = "^7.1.0"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 